+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add support for interleaving transactions from different instances of `ArezContext`. Beginning a transaction
   in one context will suspend the active transaction even if it is from a different context. The goal is to
   enable isolation between multiple contexts running within the same application.
+* **\[core\]** Introduced `ArezTestUtil.resetConfig(boolean productionMode)` to simplify test setup.
 
 ##### Fixed
 * **\[processor\]** Ensure that the types of the setter and getter for an `@Observable` property are the